Vector ID tracking added.
vec_cat: ID counter added. Counts output vectors, but
contrary to the iteration counter, it does not reset.

top_cnt1: delay shiftregisters added that delay and store
the IDs of vectors in a way, that comparator results
appear simultaneously on the comparator output with
the A and B vector IDs.